Fix a cross-cutting problem related to a bad definition of linking.
The problem was that linking of components was poorly designed, with components
already expected to already include the exports of the components they link
with, rather than them being inserted by the linking process.
This worked in principle, except that a single component could not be linked
with two different other components (as in the statement of full abstraction).
The fix is to introduce a more reasonable linking setup: components specify
imports as symbol-address pairs indicating the address for the symbol should be
substituted at the specified address in their data section. They specify exports
as symbol-word pairs, indicating the specified symbol is defined to the
specified word.
This change implied cross-cutting changes to everything related to components,
all implemented in this commit.