Added T_A to well-formedness judg., fixed comp jdg., and fixed full-abstraction pf.
T_A was added to main judgement and comp judgement.
Now the call case of the comp judgement puts no restrictions on calls outside of
ta.
The full-abstraction proof needed to be fixed with a lemma that allows the
trusted addresses in the well-formedness judgement to be increased to the ones
of the program that is plugged into it (otherwise some of the lemmas did not apply).